requirements.txt
annotated-types==0.6.0
anyio==3.7.1
cachetools==5.3.2
certifi==2023.11.17
charset-normalizer==3.3.2
colorama==0.4.6
distro==1.8.0
exceptiongroup==1.2.0
google-api-core==2.15.0
google-auth==2.26.2
google-cloud-speech==2.23.0
google-cloud-texttospeech==2.16.3
googleapis-common-protos==1.62.0
graphviz==0.20.1
grpcio==1.60.0
grpcio-status==1.60.0
h11==0.14.0
httpcore==1.0.2
httpx==0.25.2
idna==3.6
openai==1.3.5
proto-plus==1.23.0
protobuf==4.25.2
pyasn1==0.5.1
pyasn1-modules==0.3.0
PyAudio==0.2.14
pydantic==2.5.2
pydantic_core==2.14.5
pydot==1.4.2
pyparsing==3.1.1
python-dotenv==1.0.0
python-statemachine==2.1.2
requests==2.31.0
rsa==4.9
sniffio==1.3.0
tqdm==4.66.1
typing_extensions==4.8.0
urllib3==2.1.0
